+++ src/Polhem.OAuth2/Auth0/Auth0OAuth2Provider.cs
@@ -12,7 +12,7 @@ namespace Polhem.OAuth2
         /// </summary>
         /// <param name="options">The Auth0 OAuth2 options.</param>
         /// <param name="httpClientFactory">Returns the HTTP client for a request to the provider, or null to use a shared instance.</param>
-        public Auth0OAuth2Provider(Auth0OAuth2Options options, Func<HttpClient>? httpClientFactory = null) : base(options, httpClientFactory)
+        public Auth0OAuth2Provider(Auth0OAuth2Options options, Func<HttpClient>? httpClientFactory) : base(options, httpClientFactory)
         {
         }